Backtraced a bit to establish and try debug before continuing the pwm
Just this, algo some doc as a new pinout image, space cadet modification to try if it works and little more
