+++ keyboards/handwired/stm32/rules.mk
@@ -5,8 +5,6 @@ COMMAND_ENABLE = yes    # Commands for debug and configuration
 SLEEP_LED_ENABLE = yes  # Breathing sleep LED during USB suspend
 NKRO_ENABLE = yes	    # USB Nkey Rollover
 CUSTOM_MATRIX = yes 	# Custom matrix file
-#Added backlight
-BACKLIGHT_ENABLE = yes      # Enable keyboard backlight functionality
 MOUSEKEY_ENABLE = no
 
 DEFAULT_FOLDER = handwired/stm32/stm70
