--- keyboards/handwired/stm32original/README.TXT
@@ -1,9 +1,0 @@
-/*		B3	A15	  A10	A9	A8	B15	 B14  B13	B12	 B11  B10  B1	B0	A7			A6		
-	B9	ESC| F1 | F2 | F3 | F4 | F5 | F6 | F7 | F8 | F9 | F10| F11| F12| PSC| PGU| PGD
-	B8	GRA| 1  | 2  | 3  | 4  | 5  | 6  | 7  | 8  | 9  | 0  | '  | ¡  |   BCKS  | DEL
-	B7	TAB| Q  | W  | E  | R  | T  | Y  | U  | I  | O  | P  | `  | +  |         | HOM
-	B6	º	(CAP| A  | S  | D  | F  | G  | H  | J  | K  | L  | Ñ  | ´  | Ç  |   ENTER | END
-	B5	LSI| <  | Z  | X  | C  | V  | B  | N  | M  | ,  | .  | -  |   RSHIF | UP | FNL
-	B4	LCT| WIN| ALT|               ESPACE             | ALG| FN | RCT| LEF| DOW| RIG
-*/
-
