--- keyboards/handwired/stm32original/boards/GENERIC_STM32_F103/board.mk
@@ -1,5 +1,0 @@
-# List of all the board related files.
-BOARDSRC = $(BOARD_PATH)/boards/GENERIC_STM32_F103/board.c
-
-# Required include directories
-BOARDINC = $(BOARD_PATH)/boards/GENERIC_STM32_F103
