--- keyboards/handwired/stm32original/chibios_test.c
@@ -1,1 +1,0 @@
-#include "chibios_test.h"
