--- keyboards/handwired/stm32original/rules.mk
@@ -1,10 +1,0 @@
-#BOOTMAGIC_ENABLE = yes	# Virtual DIP switch configuration
-EXTRAKEY_ENABLE = yes	# Audio control and System control
-CONSOLE_ENABLE = yes	# Console for debug
-COMMAND_ENABLE = yes    # Commands for debug and configuration
-SLEEP_LED_ENABLE = yes  # Breathing sleep LED during USB suspend
-NKRO_ENABLE = yes	    # USB Nkey Rollover
-CUSTOM_MATRIX = yes 	# Custom matrix file
-MOUSEKEY_ENABLE = no
-
-DEFAULT_FOLDER = handwired/stm32original/stm32
