--- keyboards/handwired/stm32original/stm32/flash.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-Arduino_STM32_usb_hid/tools/linux/maple_upload ttyACM0 2 1EAF:0003 build/ch.bin
