--- keyboards/handwired/stm32original/stm32/stm32_f103_onekey.c
@@ -1,1 +1,0 @@
-#include "stm32_f103_onekey.h"
