--- keyboards/handwired/stm32original/stm32/stm32_f103_onekey.h
@@ -1,6 +1,0 @@
-#ifndef STM32_F103_ONEKEY_H
-#define STM32_F103_ONEKEY_H
-#include "chibios_test.h"
-//  Modified by Xydane
-//	#define LAYOUT(k00) {{ k00 }}
-#endif
